Time: 20 ms (28.46%) | Memory: 36.3 MB (15.99%) - LeetSync

+++ 689-maximum-sum-of-3-non-overlapping-subarrays/maximum-sum-of-3-non-overlapping-subarrays.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     vector<int> maxSumOfThreeSubarrays(vector<int>& nums, int k) {
         int n = nums.size();
-        vector<long long> pre(n + 1);
+        vector<int> pre(n + 1);
         for(int i = 1; i <= n; i++){
             pre[i] = pre[i - 1] + nums[i - 1];
         }
-        vector<vector<pair<long long, int>>> dp(n + 1, vector<pair<long long, int>> (3, {0, -1}));
+        vector<vector<pair<int, int>>> dp(n + 1, vector<pair<int, int>> (3, {0, -1}));
         for(int i = 0; i < 3; i++){
             dp[(i + 1)*k][i] = {pre[(i + 1)*k], i*k + 1};
             for(int j = (i + 1)*k + 1; j <= n; j++){
